+++ R/aggte.R
@@ -19,12 +19,18 @@
 #'   "calendar" (this computes average treatment effects across different
 #'   time periods; here the overall effect averages the effect across each
 #'   time period).
-#' @param balance.e If set (and if one computes dynamic effects), it balances
+#' @param balance_e If set (and if one computes dynamic effects), it balances
 #'  the sample with respect to event time.  For example, if \code{balance.e=2},
 #'  \code{aggte} will drop groups that are not exposed to treatment for
 #'  at least three periods. (the initial period when \code{e=0} as well as the
 #'  next two periods when \code{e=1} and the \code{e=2}).  This ensures that
 #'  the composition of groups does not change when event time changes.
+#' @param min_e For event studies, this is the smallest event time to compute
+#'  dynamic effects for.  By default, \code{min_e = -Inf} so that effects at
+#'  all lengths of exposure are computed.
+#' @param max_e For event studies, this is the largest event time to compute
+#'  dynamic effects for.  By default, \code{max_e = Inf} so that effects at
+#'  all lengths of exposure are computed.
 #' @param na.rm Logical value if we are to remove missing Values from analyses. Defaults is FALSE.
 #' @param bstrap Boolean for whether or not to compute standard errors using
 #'  the multiplier boostrap.  If standard errors are clustered, then one
@@ -32,6 +38,7 @@
 #'  standard errors are reported.
 #' @param biters The number of boostrap iterations to use.  The default is the value set in the MP object,
 #'  and this is only applicable if \code{bstrap=TRUE}.
+#'
 #' @param cband Boolean for whether or not to compute a uniform confidence
 #'  band that covers all of the group-time average treatment effects
 #'  with fixed probability \code{1-alp}.  In order to compute uniform confidence
@@ -48,7 +55,9 @@
 #' @export
 aggte <- function(MP,
                   type = "group",
-                  balance.e = NULL,
+                  balance_e = NULL,
+                  min_e = -Inf,
+                  max_e = Inf,
                   na.rm = FALSE,
                   bstrap = NULL,
                   biters = NULL,
@@ -56,13 +65,15 @@ aggte <- function(MP,
                   alp = NULL,
                   clustervars = NULL
                   ) {
-  compute.aggte(MP,
-                type,
-                balance.e,
-                na.rm,
-                bstrap,
-                biters,
-                cband,
-                alp,
-                clustervars)
+  compute.aggte(MP = MP,
+                type = type,
+                balance_e = balance_e,
+                min_e = min_e,
+                max_e = max_e, 
+                na.rm = na.rm,
+                bstrap = bstrap,
+                biters = biters,
+                cband = cband,
+                alp = alp,
+                clustervars = clustervars)
 }